requirements.txt
python-telegram-bot==20.7
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.2.1
pydantic==2.12.4
openai==1.14.0

